data_store/requirements.txt
connexion==2.5.1
swagger-ui==0.1.2
SQLAlchemy==1.3.13
mysql-connector-python==8.0.19
PyMySQL==0.9.3
pykafka==2.8.0
Flask==1.1.1
Flask-Cors==3.0.8